daemon/c_net: Make old API for netif moving/removing private
We now use the new API for adding and removing physical interfaces
in ol subsystems of the cmld. Thus, we can make the old functions
private again. The c_net_move_ifi and c_net_remove_ifi functions were
not meant to be available outside of c_net.c anyway.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>