daemon/guestos_mgr: Proper image download and failed download handling
Since the connectivity change handling was removed lately, we do not
trigger downloads based on online connectivity anymore. We now trigger
downloading of GuestOS image files directly on push_guestos_config.
Further, in cases where the download fails, we purge the failed GuestOS
to allow a new update run with the same os version again.

If still it should be necessary to re download images for an existing
GuestOS version, this is possible now by a push_guestos_config with the
same version with the restriction that the already installed config is
used for integrity checks of the image files.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>