daemon/guestos_mgr: control message in case of download not started
Instead of just dropping out by return, we jump to the out label now.
This ensures that a response is given to the control client which
issued the push_guestos_config command. This code path is triggered
if two control sessions trigger a guestos update of the same version
simultaneously.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>