common/event: fixed inotify event registering
event_reset() in child processes of cmld caused the removal of all
inotify watches on the corresponding static fd. We fixed this by just
deleting the inotify handlers from the internal list as well as
cleaning up the static fd. Further, a mem leak of the corresponding io
event was fixed by defining that one static, and free it also in
event_reset();

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>