daemon/audit: handle corrupt records in audit log file
If the protobuf parsing of a record from the corresponding log file
fails, we now create a new audit record with the corrupt message as
raw_text meta value. To fixup corrupted records, we check the
delimiter of the previous record before writing any log record to the
log file.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>