daemon/uevent: Fixed masking of synth_uuid of coldboot events
During container start cold boot uevents are triggered for containers
only by setting the UUID as synth_uuid in the event. This is masked
from the event before forwarding. However we accidentally forwarded
the original event instead of the newly created one without the
synth_uuid. This is herby fixed!

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>