daemon/smartcard: disable certificate time check for provisioning
During provisioning the real time clock needs to be set to a
proper value. For that the c0 needs to bring up network and sets
the clock properly once. Therefore, the validity check needs to
be ignored during provisioning until the clock is written to the
hw_clock. Not till then certificate validation makes sense.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>