daemon/c_vol: Fixed mapping of '/dev/console' in pre_exec hook
Lately, the tty console mapping for systemd was introduced during
population of the container's '/dev' tree in the pre_exec hook.
However, we missed to initialize a pointer to NULL there, which may
lead to a protection fault. Further, the static boolean in the
callback function was wrong. This lead to that only one container
have the mapping. We now use a structure as callback data which
holds the values for tty name and the bool which decides if a
mapping should be made or not.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>