daemon/c_net: Properly release offsets for ip and interface names
The cleanup of network interfaces was moved to a separate child,
allowing to cleanup interface endpoints in c0's netns.
However, the offsets for ip and network veth naming were not released
in the main cmld process. We plit out the interface destruction from
c_net_cleanup_interface to a new function c_net_interface_down. This
will be executed in the child process while we moved the execution of
c_net_cleanup_interface to the c_net_cleanup method of the main process.

Signed-off-by: Michael Weiß <weiss@m4st3rnet.de>