daemon: Integrate usbtoken in CML uevent subsystem
Integrating the usbtoken functionaility in the existing
uevent subsystem allows for better hotplugging support.
When a usbtoken reader is detached the associated container
is immediately terminated.
When a reader is attached to a running system on which a
container that is associated with the reader is present,
the token is being initialized inside the scd.

Signed-off-by: Christian <christian.epple@aisec.fraunhofer.de>