daemon/audit: reflect changes of mainline dm-audit kernel module
We switch to the mainline kernel module dm-audit, which uses
AUDIT_DM_CTRL (1338) and AUDIT_DM_EVENT (1339) events to log audit
messages of the device-mapper (integrity and crypt) devices.
We handle integrity violations as part of AUDIT_DM_EVENT directly,
and forward AUDIT_CM_CTRL messages to the corresponding container.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>