daemon: properly cleanup with 'atexit' of stdlib
Instead of registering an on exit helper function in main submodule,
we now use the atexit() feature of stdlib to register relevant cleanup
methods at the point in code were the corresponding init functions are
called. This provides proper cleanup, whenever an exit() is called in
the main process of the cmld. Any forked helper childs of cmld in the
corresponding submodules have to call _exit() instead of exit() now,
since the atexit registered methods survive the fork.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>