daemon/c_net: properly kill udhcpd server on interface down
Since the pid of the corresponding dhcpd server is only known in the
helper child which is destroyed after initial configuration, the pid
is not set in the ni structure when a new interface helper for
cleanup in c0 is spawned. Thus we relay on the pid stored in th pid
file in the file system on cleanup.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>