daemon/container, c_run: port c_run submodule to new module facility
Corresponding direct API which is exported by container.h was hidden
in function pointers which can be registered by the submodule.
Thus, a c_run.h is not necessary anymore. With this commit is now
possible to remove the c_run module (c_run.c) from Makefile and
everything works transparently without abilitiy to run commands.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>