daemon/container, c_audit: port c_audit submodule to new module facility
Corresponding direct API which is exported by container.h was hidden
in function pointers which can be registered by the submodule.
Thus, a c_audit.h is not necessary anymore. With this commit is now
possible to remove the c_audit module (c_audit.c) from Makefile and
everything works transparently without abilitiy to signal audit events
to containers.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>