common/uevent: update end of uevent after member replacement
After updating a member with a smaller size as before, the new
uevent was not terminated by a '\0' char. This was leading to parse
till the end of the uevent buffer reading chunks of the previous event.
We fixed this by just writing the missing '\0' char at msg->len in
the buffer. To avoid buffer overflow the maximum buffer size is
checked in advance.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>