common/uuid: allow to use a hex string as initialization for uuid
Instead of only accepting a uuid formatted string such as
8aff42d9-189e-47e7-9f1c-6c6ff41d8ee3 we also accept a hex string
of arbitrary length grater then 128bit such as
8aff42d9189e47e79f1c-6c6ff41d8ee3ffffffffffffffffffffffffff
as input for the uuid. The resulting uuid of course is truncated
to the first 128bit.
This can be used to accept docker ids which are 256 bit random hex
strings as input for a container id.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>