daemon/oci, c_oci: Initial module for oci compatible containers
This module implements a protobuf control interface (oci_control)
which receives commands from an external oci wrapper binary (rung,
see next commit). The runtme binary implements a oci runtime command
line interface and wraps those commands to oci_control messages.
Including to transfer the json-based con fig file. The oci module makes
use of the libcoispec to parse that config and extracts relevant parts
which are mapped to a cml container. 'c_oci' implements the corresponding
compartment module which pauses the counter start during oci create.
The oci CREATE command should start the namespace but not the child
process, those this corresponds to the cml contaienr state BOOTING.
At that point we use the pause() syscall to pause execution of the
container child just before execve of init. After a oci START command
is received the oci module sends an interrupt to the child process of
the container and continues execution.

Note that this is module only provide rudimentary oci complainant
functionality and does not yet fulfil the oci spec completely.
Especially, hooks are not yet implemented. Further, we enfore safe
defaults, hard set namespaces, cgroups and capabilities and do ignore
these parts of the oci josn-config. Further, do we implement a
writable 'confidential' layer over the oci rootfs which is only
accessible in the CML's mount namespace.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>