daemon/oci, cmld: map oci_control socket to containers
In case of feature_install_guest is activated or c0, the container
gets the listening socket bound inside. Thus, it is possible to run
'rung' as oci container runtime inside a container. With this approach
it is possible to run docker/containerd inside an unprivileged user
namespaced container, while the cmld will execute the corresponding
application container as a separate container. This allows us to hide
user activity as well as user data from the corresponding docker or
containerd instance.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>