daemon/c_user, container: new open_userns() handler
We provide a new container handler, open_userns(), which is implemented
in c_user submodule. This handler starts a new dummy child in a userns
with the containers uid/gid mappings and returns a file descriptor to
that new user namespace. This could be used for idmapped mounts later on.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>