daemon/c_idmapped: New module for idmapped mounts to replace c_shiftid
The c_idmapped submodule provides id shifting based on the mainline
idmapped mount feature. It needs also the in Linux v6.0 introduced overlay
mounting of shifted layers. This module can be used as replacement for
c_shiftid which is based on non-mainline shiftfs. Also a fallback with
bind mounts and chowing files for older kernels is included (ported from
c_shiftfs) in c_idmapped. Note that tmpfs does not support id shifting,
as well as in user namespace mounted file systems won't accept device nodes.
Thus, tmpfs for dev is mounted in root ns and bind mounted to /dev.
tmpfs used for overlays are just over mounted in userns with a new tmpfs.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>