daemon/c_cgroups_v2: properly run cmld in cgroup leaf
For v2 cgroups, the rule no inner process has to be fulfilled. Thus,
we have to move the cmld itself into a leaf cgroup. For this, we make
use of the previously introduced common helper, to identify in which
cgroup the cmld is running in. A new subgroup for the cmld itself
is created in this cgroup.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>