daemon/c_cgroups_systemd: Submodule to provide systemd compliance
Systemd proclaims the rule of a single manger for cgroups subtrees.
Thus, if we are running as daemon/service on a host using systemd as
init system, this module implements the corresponding api to delegate
a subtree to the cmld. That way, we leave the root cgroup untouched
and completely manged by systemd.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>