rattestation/attestation: Correct portable format string for size_t
The printed variables are of type size_t. size_t is a typedef for long
unsigned int on 64bit architectures and unsigned int on 32bit architectures.
This commit fixes the format string to correctly recognize the size of type
size_t on all architectures and not always assume 64bit. This is required to
fix compilation for the arm32 (e.g. colibri-imx6ull board) architecture.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>