daemon/c_smartcard: fixed size check of wrapped key
In case the wrapped key file is a file of size 0, the resulting
protobuf message triggers an assertion in scd. It makes sense to
check if the size of the wrapped key which was read has the
proper TOKEN_KEY_LEN. This would also lead to correctly trigger
the corresponding audit event.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>