daemon/container: fixed killing whole user space on container_kill
In case a container_kill is triggered for a container object which
has the the pid not set to -1, e.g. a container in state ZOMBIE,
the whole user space was killed.  A check for pid == -1 was missing
before giving that to kill() syscall. This causes a sig kill to all
processes but init. The proper check is now added.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>