daemon/c_cgroups_v2: add an extra level of hierarchy for container
Since systemd "insist on managing the delegated cgroup tree's
top-level attributes" [https://systemd.io/CGROUP_DELEGATION/], it
is necessary to delegate a subtree of the container subtree
which is managed by the cmld. Thus, we introduce a child cgroup.
The resource limits are enforced in the container specific upper
level of that cgroup. Thus, if we run systemd inside the container
it only gets access to the child cgroup and can safely manipulate
the top-level attributes of that cgroup.

The following example shows the hierarchy, cmld manages the container
subtree "f1b99085-fcf8-470d-96c7-6c975ed1db55" and the systemd inside
the "init.scope" can manipulate its own subtree below and including
"child":

  ├─gyroidos.scope …
  │ ├─f1b99085-fcf8-470d-96c7-6c975ed1db55
  │ │ └─child
  │ │   ├─init.scope
  │ │   │ └─111869 /sbin/init
  │ │   └─system.slice
  │ │     ├─systemd-udevd.service
  │ │     │ └─111928 /lib/systemd/systemd-udevd
  │ │     └─...
  │ └─cmld
  │   ├─111820 /sbin/cmld
  │   └─111879 udhcpd -f /run/udhcpd/r_0.conf

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>