daemon/c_user: Release userns bound to file system properly on stop
At the time when c_user_cleanup() is called, the user namespace is
still in use, we have to release the bound to the user namespace in
the cmlds file system after all other cleanup() methods were called
to. Thus, we register an observer to the internal state machine
which executes the release operations after a container entered
the STOPPED state.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>