daemon/c_idmapped: Properly release temporary bind mounts
In IDMAPPED_SRC_DIR several temporary mount for idmapping are
created during startup of a container. Those mounts are cleaned
up in c_idmapped_umount_dir_cb(). The directory is recursively
checked for mount points and if so umounted. However, a mount
point is used twice sometimes by the idmapped module, e.g., when
over-mounting read-only file system with a writable tmpfs. We
switched from if to while a file is a mount point in the umount
callback c_idmapped_umount_dir_cb() to properly release those
mounts.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>