daemon/c_shiftid: properly allocate mnt in c_shiftid_mount_shifted()
In case of ovl mounts where the lower dir should be mounted shifted
and target is not the same directory, the mnt struct was not allocated
but used for setting the mnt->target. This caused a NULL pointer
dereference. This is fixed now by allocating mnt early in
c_shiftid_mount_shifted().

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>