daemon/c_vol: use file_links_to_blk() for proper reboots
During reboot the device mapper devices are kept alive.However, the
check for the corresponding block device during restart was failing
since usually the used path "/dev/mapper/<device-label>" is a symlink.
We now use the file_links_to_blk() helper to check if the path links
to the corresponding block device. This allows to reuse the existing
block device during reboots.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>