daemon/c_cgroups_v2: respect intermediate child cgroup in add_pid
In the 'start_pre_exec' hook we apply another indirection (subcgroup
for container child). This is at least needed to run systemd as
container payload. We have to consider this intermediate cgroup in
the 'add_pid' function, too. This, is called, e.g., by the c_run
submodule to join the container's cgroup.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>