daemon/uevent: fixed renaming of devpath in hotplug_rename_interface()
uevent_rename_interface() the renamed devpath object was made on the
original event instead of the one with the renamed ifname.
Further, it was ignored anyway since the uev_chname was returned and
not the uev_chdevname. We fixed that and additionally provide proper
cleanup of memory and error handling.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>
(backport from 130a33bc5a2025492b944291f9ffe610d019419c)