daemon/c_shiftid: removed overlayfs mounting in userns
The code to mount overlayfs inside of userns was buggy and also
not properly working for all of our use cases. Thus, we just removed
that and now use the properly working code which was introduced for
backward compatibility for kernels which did not support overlayfs
mounts in userns anyway.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>