daemon/c_idmapped: revert previously applied mapping by chmod
In case, a container was created on a system without proper
idmapped mount support, the files are owend by uid of the user
namespaced root uid. Thus, we have to revert this by chowning
back to system root uid. This is done by generalizing the
c_idmapped_chown_dir_cb() to use uid and gid as callback data
and do not deppend on the container uid. Thus, it is now
feasible to call the helper with uid, gid = container_uid to
apply the mapping. And uid, gid = 0 to revert it.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>