daemon/c_vol: activate idshifting for MOUNT_TYPE_SHARED
During image mounting on images with MOUNT_TYPE_SHARED shifting
of ids was disabled. The code to allow mounting read-only file
systems even if neither siftfs nor idmapped mounts are supported,
is in place of the corresponding modules c_shiftid and c_idmapped.
Thus, we now can allow this here.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>