common/proc: fixed error in waitpid causing busy loop
To ignore signals caused by interrupted syscall, waitpid() was
called non blocking in a loop until the pid of the watched child
was returned. However, this is causing a busy loop. Thus, we now
run blocking again and check the errno against EINT to stay in
the loop until the pid of watched child really terminates.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>