scripts/ci/VM-container-tests: retrieve logs by e2tools
Do not depend on running VM and core0 container anymore to retrieve
logs. We just extract the data partion of the raw disk image with dd.
The logs are in /userdata/logs there. With e2ls and e2cp we can
iteratively get the logs from the raw image after the testing VM has
been stopped. This should work even in error cases.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>