daemon/guestos: Download image files containing the verity hash tree
Introduce a second stage of callback indirection on images which need
a verity hash tree. First we check if a verity_sha256 is set on the mount
entry and download the corresponding image. In the cb the former image
download callback is than triggered. In case no verity hash is set for
the entry directly use the former download callback.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>