daemon: Introduce compartment flags for feature configuration
Now the compartment type is hold as flag as well as the namespace
configuration. The previous compartment_type_t is now specified
as container_type_t to transparency map the existing container API.
The compartment flags attribute is currently hidden in compartment_t
and only exported to the higher level container module.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>