daemon/compartment: Introduce new start hook
This commit introduces a new hook during compartment start hook called
start_pre_exec_child_early. It is executed in the child process at compartment
start prior to the start_pre_exec hook in the parent process. The new hook
is needed by the c_seccomp module to register the seccomp filter obtaining the
notify fd to be fetched by the parent process.

Signed-off-by: Felix Wruck <felix.wruck@aisec.fraunhofer.de>