daemon/c_idmapped: add a mnt struct for ovl upper image mounts
To get the upper directory mapped correctly we need to add a mnt
struct as mapping for the corresponding image. Otherwise the upper
directory would just be created somewhere in the cml storage. This
was resulting in exhausting the ram since "persistent" data of the
containers where written to /tmp/ of the CML which is a tmpfs.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>