daemon/c_net: Fixed a missing NULL pointer check in add_interface()
The if_name may be NULL if it is missing in configuration or a mac
address is specified in the configuration to which the interface is
not present while calling add_interface().
Now just return -1 as error in that case and do not call
c_net_move_ifi() which would assert if_name.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>