daemon/c_net: do not free pnet_cfg on error in c_net_add_interface()
In error cases of c_net_add_interface() in the c_net submodule, the
pnet_cfg struct which was allocated out side was freed. This have
lead to an use-after-free in some error cases during compartment
start. Especially, when the compartment was started again after an
error, this may have lead to a segfault in cmld's main process. Thus,
do not free this structure in the c_net submodule anymore and handle
this in the calling functions in the hotplug as well as the control
module.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>