daemon/c_cgroups_dev: Properly free data of allowed dev list
The allowed list is used to cleanup the global allowed list about.
After that accidentally the allowed list it self was deleted.
The loop to cleanup the data of the allowed list elements itself
was deactivated that way because the list itself was already NULL.
This is fixed now by just remove the wrongly placed list deletion
code.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>