daemon/c_cgroups_dev: added missing free of assign list
dev_items are added by c_cgroups_dev_list_add() to assigned_devs
and allowed_devs. This adds a deep copy of the dev_item to those
lists each. Therefor, both lists needs to free the corresponding
copy on cleanup(). Further, c_cgroups_dev_deny() also removes
just the dev_item from the lists but did not free'ed it.
This is both fixed in this commit.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>