daemon/compartment: Check compartment pid/early_pid in sigchild_cb
If the container gets terminated the sigchld handler calls cleanup.
The compartment_cleanup() sets internal the compartment->pid and
compartment->early_pid to "-1". In case, another SIGCHLD arrives,
e.g., c_fifo kills its forwarding children, the sigchld handler for
the compartment is triggered again. This lead to a waidpid() on 1
since we do a waitpid() for the process group by -(compartment->pid).
This resulted in a second cleanup run. To prevent this we just
check compartment->pid to be a valid pid before running waitpaid()
on it.

This fixes commit "(1aa1055fb09f) daemon/compartment: Cleanup and
handle helper children properly"

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>