common/nl: add nl_sock routing function for IFADDR events
This provides the ability to listen to ip address changes on network
interfaces. Removed the nl_sock_conf_xfrm_sock() function because it does
not add anything except the nl_groups. Did not add the groups parameter to
the nl_sock_conf_uevent_sock() function because it should subscribe to all
kernel messages.

Signed-off-by: David Sonntag <dsonntag@blackned.de>