scd/usbtoken: keep track of used card terminals
If multiple tokens are plugged and unplugged during runtime it is
not sufficient to just use a global counter for card terminal
numbers. Additionally to g_ctn we now keep track of already used
ctns in a list. If a new token is generated (plugged) we either
reuse one of the ctns of the list or just increment the g_ctn
and use that. On token remove (unplug) we set the ctn as available
by added it to the list and do not decrement the g_ctn anymore.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>