common/event: Do not add the same inotify event again
In event_add_inotify(), if the event_inotify_t is already in list,
do not add it again and return -EEXIST. Thus, double triggering the
callback would be avoided as well as an acidentialy triggered double
free in event_reset().

This also fixes errors on reboot, since the inotify event for
auto mounting in c_uevent is added in a container start hook and
removed in the container stop hook. The stop hook is not triggered
on container internal reboots. In that case the event remains in
the internal list and would be added again on startup, resulting
in a duplicat in the inotify list of event.c. The cloned early child
runs event_reset() which clears the list and frees any element
in the list. When it comes to the duplicate the data was allready
freed resulting in a double free.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>