daemon/c_user: only keep userns reference during reboot
Do not bin userns to file system if not necessary. We do not need
to keep userns alive during normal operation. The mechanism to keep
it alive if the last container process died is only needed while
rebooting of a container. Thus, remove the observer callback and
just bind the userns in clenup if rebooting. When the container is
restarted again just remove the bound and close the fd.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>