daemon/download: Use -1 instead of file_size() for file_copy()
If downloading images locally by file_copy() just use -1 as count
parameter instead of file_size() of the source file. Since the
parameter does not take the size for the output file but the count
of output blocks. The output file size would then be count * bs.

Further, we restructure the method a little to get better log
output if we are using wget or internal file_copy().

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>