daemon/c_user: Keep userns active by opening nsfd only
Instead of binding nsfd to the file system, we just open the
corresponding nsfd from procfs at start hook. This fd is closed
in the stop hook. During reboots the fd is stored in the user
struct and can be joined by setns directly.

The used ns_bind() method which was use untill now, did not
work since the container child process does not exist in procfs
any more when cleanup method is called on reboots.

This fixes following error on reboot of c0:
 <ERROR> ns.c+337: Could not bind mount ns file /proc/597/ns/user on
	/var/run/userns/00000000-0000-0000-0000-000000000000 (2: No such file or directory)
 <WARN>  c_user.c+261: Could not bind userns of core0 into filesystem!

Fixes: a1e35a648a80 ("daemon/c_user: only keep userns reference during reboot")
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>