daemon/c_seccomp: Refactor BPF filter for multiarch support
Refactor the BPF filter so that the allowed architecture is determined
at compile time and not hard-coded to x86. Only insert the
X32_SYSCALL_BIT check when compiling for x86. The new filter can be
extended (if required) to support secondary architectures in
compatibility mode (x86_64 -> i386, aarch64 -> arm). At the moment
syscalls using the compatibility architecture are all blocked and
processes are killed.

We do not implement a second architecture at the moment due to the lack
of a dependable source of secondary architecture syscall numbers for all
architectures.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>