daemon/{guestos,main,hardware}: Remove/move hardware_get_name
hardware_get_name usage in main.c is purely informational an can be
removed.

hardware_get_name is therefore only used in guestos.c where its
implementation is moved to. To determine the machine at runtime, the
gyroidos machine string is determined by the Linux kernel machine from
uname.

As this is the last function in hardware.h the file is removed along
with all includes.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>