daemon/c_net: allow to set name of veth endpoint in CML/c0
Currently, the 'if_rootns_name' of ContainerVnetConfig is used as
read-only field to show the cml's chosen if name at container start.
However, users where confused why this was not read from config and
just overwritten with the runtime value.

We now just take the value of 'if_rootns_name' and set it as name
for the veth endpoint in the rootns (c0) instead of the 'r_*'
values. However, the generated 'r_*' is taken as fallback if the
user provided name in config does not work.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>