daemon/c_net: Cleanup veth endpoints in rootns by ifindex
If the veth endpoint in rootns was renamed during runtime of a
container, it could not be cleaned up properly in the cmld, due
to the use of the name which was known at startup.

Thus, we introduce a new field to 'veth_cmld_idx' in the struct
'c_net_interface_t' to store the index before the container is
started. We use this index during cleanup as this does not change
if the interface is renamed.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>