common/sock: derive env_var name by sock_name in sock_get_path_new
In sock_get_path_new(), the sock_name was used directly as variable
name in the environment. This however is not conform to variable
name convention in the environment. We now use the sock_name as
input to derive the correct variable name. "-" is replaced by "_"
and all other characters are capitalized.

The corresponding systemd unit file needs to set the environment
variable now for the cml control socket as 'CML_CONTROL'.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>