daemon/compartment: Introduce new flag COMPARTMENT_FLAG_MODULE_LOAD
The new flag COMPARTMENT_FLAG_MODULE_LOAD will be used in followup
commits to emulated module loading for those containers which have
this flag set.

Only allow this flag to make it to the internal compartment struct if
module signatures are enforced in the running kernel. We use the new
helper compartment_is_module_sig_enforced(void) to check the
proc file "sys/module/module/parameters/sig_enforce" which says "Y"
in that case.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>