daemon/c_seccomp: check capability during emulation of syscalls
We missed to check the capability of the running process in respect
to its user namesapce. Thus it may have been possible for any
processes in the container to make the cmld execute time related
syscalls for them.

We now introduce the helper c_seccomp_capable() which checks if the
corresponding capability is in the effective set of the corresponding
process. This, helper uses the output of /proc/<pid>/status.

All currently implemented time related syscalls are now checked
before emulation.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>