daemon/c_seccomp: handle mknod() and mknodat()
We only do user space emulation for block and char devices, otherwise
the bpf-based secompfilter directly jumps to SECCOMP_RET_ALLOW and let
the kernel process the syscall.

We emulate mknod() by mknodat() for code dedup. mknod() and mknodat()
have the same arguments, except that mknodat has dirfd as first argument:

  int mknod(const char *pathname, mode_t mode, dev_t dev);
  int mknodat(int dirfd, const char *pathname, mode_t mode, dev_t dev);

We therefore use an offset +1 for the data.args array to retrieve syscall
parameters from seccomp req in case of mknodat() and set dirfd to AT_FDCWD
in case of mknod().

Further, we check CAP_MKNOD of the container process and if the c_cgroups
submodule of the container allows the device access.

The actual syscall execution is made in a forked child which joins the
mount namespace of the container. We also change to the mapped root user
uid. Otherwise, if we just use system root with uid 0, we cannot write
mounts mounted by the container itself, e.g. '/tmp'. This would result in
an "errno (75: Value too large for defined data type)".
To still allow mknod we need to preserve CAP_MKNOD for the forked child.

Further, we need to switch to the current working directory of the
target process for which we do the emulation. proc_get_cwd_new() is used
to retrieve the CWD before switching namespaces. In the forked child,
just a chdir() to that directory is made then.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>