daemon/c_seccomp: check allowed list during finit_module() emulation
In c_seccomp_new(), use the allowed list from the container object
(config) to retrieve a list of modules according the dependencies in
'modules.dep' file and store this in the c_seccomp struct.

Later during emulation check if the module name is in the generated
list of allowed modules and their depending modules, if not drop out
of emulation early.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>