daemon/guestos: Relative download paths
This commit allows the update_base_url parameter in the device or
guestos config to take on a relative format without leading /. In this
case the update url is interpreted relative to the calling containers
root directory which is accessed via /proc/<pid>/root. The behaviour of
absolute update paths with leading / is unchanged.

Signed-off-by: Johannes Wiesböck <johannes.wiesboeck@aisec.fraunhofer.de>