daemon/hotplug: trigger uevents for missing net devices on early boot
The current procedure to activate all devices in the CML init script
and then added all available network interfaces to the cmld's internal
tracking list, is racy.

E.g., the modprobe for a driver will be made by 'init' which will
create the corresponding net device. However there is some time
needed until the corresponding sysfs structures are populated.
The cmld gets started and initializes its internal list of net devices
on basis if '/sys/class/net/<ifname>' is present. This is not always
ready at that time. The result is that the corresponding net interface
will not be added to the internal tracking list and stays in the init
root netns forever.

Thus, to avoid missing some interface to get moved to c0 on startup,
let hotplug_init() trigger all relevant uevents again.
That way the hotplug module will handle those devices as just hot
plugged ones. The mechanism in place uses a timed callback for the
sysfs to be populated.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>