daemon/c_automount: use namespace_exec to mount external block devices
Mount propagation does not work with c_idmapped submodule. Since, we
overmount the '/tmp/<container uuid>/media/' directory in the rootns with
the idmapped mount for the container, access from the root user namespace
will result in an "errno (75: Value too large for defined data type)".

We now join the mount namespace of the container and switch to the
root uid keeping System wide CAP_SYS_ADMIN to execute the mount.

Fixes following error:
 [530] <ERROR> dir.c+96: Could not mkdir /tmp/00000000-0000-0000-0000-000000000000/media/external
	 (75: Value too large for data type)

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>