common/ns: allow to keep system-wide cap with namespace_exec
Changed paramters become_root to a uid and cap integer. The previous
behaviour of become root is now acchived by setting uid to 0 and the
CLONE_NEWUSER flag must be set.

If we want to preserve system-wide cappability but join a container
namespace accting as container root user, we now can achive that to
set uid to the mapped root uid and cap to the capability we need to
retain.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>