daemon/control: allow GET_LAST_LOG in cc_mode (EXPERIMENTAL)
WARNING!! EXPERIMENTAL !! Logfile encryption is not yet implemented !!

Since CML logfiles may contain sensitive data, this command needs
special protection. The logfiles need to be encrypted by the cmld to
safely retrieve logfiles!

WARNING!! EXPERIMENTAL !! Logfile encryption is not yet implemented !!

Therefore, to enable this feature CC_MODE_EXPERIMENTAL needs to be set
as build flag to "y". The default value in the Makefile is set to "n".

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>