daemon/device_id: Introduce new module device_id for device uuid
To later allow raw updates of the device.conf, we need to maintain
the device uuid in a separate file. This module implements a wrapper
to the corresponding storage file only containing the uuid.

In the protobuf definition in device.proto, we define a new message
'DeviceId' containing the field uuid. The corresponding uuid field of
the message 'DeviceConfig' is removed and defined as reserved.

Subsequent patches will us this as replacement for the device
uuid of the device_config module.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>