daemon/cmld: Get device uuid from device_id module
In the cmld module we now retrieve the uuid from the new device_id
module instead of the device_config module.
Similar change needs to be applied to the scd, which takes care of
provisioning of the device_id.

Also remove the generation of the device uuid in device.conf of
install target in Makefile_lsb.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>